requirements.txt
black==24.10.0
pre_commit==4.0.1
pylint==3.3.1
pytest==8.3.3
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
psycopg2-binary==2.9.10
bs4==0.0.2
jsonschema==24.2.0
